backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
sentence-transformers
transformers
asyncpg
sqlalchemy
watchdog
httpx
transformers
torch
accelerate
safetensors
Pillow
torchvision